@@ -1,37 +1,36 @@
 {
-  "name": "angular-cli",
-  "version": "1.0.0-beta.24",
+  "name": "@angular/cli",
+  "version": "1.3.0",
   "description": "CLI tool for Angular",
-  "main": "packages/angular-cli/lib/cli/index.js",
+  "main": "packages/@angular/cli/lib/cli/index.js",
   "trackingCode": "UA-8594346-19",
   "bin": {
     "ng": "./bin/ng"
   },
   "keywords": [],
   "scripts": {
-    "build": "node ./scripts/publish/build.js",
-    "build:patch": "node ./scripts/patch.js",
-    "build:packages": "for PKG in packages/*; do echo Building $PKG...; tsc -p $PKG; done",
-    "test": "npm-run-all -c test:packages test:cli test:deps",
+    "build": "node scripts/run-tool.js publish build",
+    "test": "npm-run-all -c test:packages test:cli test:deps test:licenses test:messages",
     "e2e": "npm run test:e2e",
-    "e2e:nightly": "node tests/e2e_runner.js --nightly",
-    "mobile_test": "mocha tests/e2e/e2e_workflow.spec.js",
-    "test:e2e": "node tests/e2e_runner.js",
+    "e2e:nightly": "node tests/run_e2e.js --nightly",
+    "test:e2e": "node tests/run_e2e.js",
     "test:cli": "node tests/runner",
     "test:deps": "node scripts/publish/validate_dependencies.js",
     "test:inspect": "node --inspect --debug-brk tests/runner",
+    "test:licenses": "node scripts/test-licenses.js",
+    "test:messages": "node scripts/test-commit-messages.js",
     "test:packages": "node scripts/run-packages-spec.js",
-    "build-config-interface": "dtsgen packages/angular-cli/lib/config/schema.json --out packages/angular-cli/lib/config/schema.d.ts",
     "eslint": "eslint .",
-    "tslint": "tslint \"**/*.ts\" -c tslint.json -e \"**/blueprints/*/files/**/*.ts\" -e \"node_modules/**\" -e \"tmp/**\" -e \"dist/**\"",
-    "lint": "npm-run-all -c eslint tslint"
+    "tslint": "tslint \"**/*.ts\" -c tslint.json -e \"**/config/schema.d.ts\" -e \"**/tests/**\" -e \"**/blueprints/*/files/**/*.ts\" -e \"node_modules/**\" -e \"tmp/**\" -e \"dist/**\"",
+    "lint": "npm-run-all -c eslint tslint",
+    "tool": "node scripts/run-tool.js"
   },
   "repository": {
     "type": "git",
     "url": "https://github.com/angular/angular-cli.git"
   },
   "engines": {
-    "node": ">= 4.1.0",
+    "node": ">= 6.9.0",
     "npm": ">= 3.0.0"
   },
   "author": "Angular Authors",
@@ -41,127 +40,111 @@
   },
   "homepage": "https://github.com/angular/angular-cli",
   "dependencies": {
-    "@angular/compiler": "^2.3.1",
-    "@angular/compiler-cli": "^2.3.1",
-    "@angular/core": "^2.3.1",
-    "@angular/tsc-wrapped": "^0.5.0",
-    "async": "^2.1.4",
+    "@angular-devkit/build-optimizer": "0.0.13",
     "autoprefixer": "^6.5.3",
-    "chalk": "^1.1.3",
+    "chalk": "^2.0.1",
+    "circular-dependency-plugin": "^3.0.0",
     "common-tags": "^1.3.1",
-    "css-loader": "^0.23.1",
-    "debug": "^2.1.3",
+    "core-object": "^3.1.0",
+    "css-loader": "^0.28.1",
+    "cssnano": "^3.10.0",
     "denodeify": "^1.2.1",
-    "diff": "^2.2.2",
+    "diff": "^3.1.0",
     "ember-cli-normalize-entity-name": "^1.0.0",
     "ember-cli-string-utils": "^1.0.0",
-    "enhanced-resolve": "^2.3.0",
-    "exists-sync": "0.0.3",
-    "extract-text-webpack-plugin": "^2.0.0-beta.4",
-    "file-loader": "^0.8.5",
-    "findup": "0.1.5",
-    "fs-extra": "^0.30.0",
+    "enhanced-resolve": "^3.4.1",
+    "exports-loader": "^0.6.3",
+    "extract-text-webpack-plugin": "3.0.0",
+    "file-loader": "^0.10.0",
+    "fs-extra": "^4.0.0",
     "get-caller-file": "^1.0.0",
-    "git-repo-info": "^1.0.4",
     "glob": "^7.0.3",
-    "html-webpack-plugin": "^2.19.0",
+    "heimdalljs": "^0.2.4",
+    "heimdalljs-logger": "^0.1.9",
+    "html-webpack-plugin": "^2.29.0",
     "inflection": "^1.7.0",
-    "inquirer": "^0.12.0",
-    "is-git-url": "^0.2.0",
-    "isbinaryfile": "^2.0.3",
-    "json-loader": "^0.5.4",
-    "karma-sourcemap-loader": "^0.3.7",
-    "karma-webpack": "^1.8.0",
-    "less": "^2.7.1",
-    "less-loader": "^2.2.3",
-    "loader-utils": "^0.2.16",
+    "inquirer": "^3.0.0",
+    "isbinaryfile": "^3.0.0",
+    "istanbul-instrumenter-loader": "^2.0.0",
+    "karma-source-map-support": "^1.2.0",
+    "less": "^2.7.2",
+    "less-loader": "^4.0.5",
+    "license-webpack-plugin": "^0.5.1",
+    "loader-utils": "^1.0.2",
     "lodash": "^4.11.1",
-    "magic-string": "^0.16.0",
-    "markdown-it": "4.3.0",
-    "markdown-it-terminal": "0.0.3",
+    "magic-string": "^0.22.3",
+    "memory-fs": "^0.4.1",
     "minimatch": "^3.0.3",
     "node-modules-path": "^1.0.0",
-    "node-sass": "^3.10.1",
-    "nopt": "^3.0.1",
-    "offline-plugin": "^3.4.1",
-    "opn": "4.0.1",
-    "ora": "^0.2.0",
-    "portfinder": "1.0.9",
-    "postcss-discard-comments": "^2.0.4",
-    "postcss-loader": "^0.9.1",
-    "quick-temp": "0.1.5",
+    "nopt": "^4.0.1",
+    "opn": "~5.1.0",
+    "portfinder": "~1.0.12",
+    "postcss-loader": "^1.3.3",
+    "postcss-url": "^5.1.2",
     "raw-loader": "^0.5.1",
-    "readline2": "0.1.1",
-    "reflect-metadata": "^0.1.8",
-    "remap-istanbul": "^0.6.4",
     "resolve": "^1.1.7",
-    "rimraf": "^2.5.3",
     "rsvp": "^3.0.17",
-    "rxjs": "^5.0.1",
-    "sass-loader": "^4.0.1",
+    "rxjs": "^5.4.2",
+    "sass-loader": "^6.0.3",
     "script-loader": "^0.7.0",
     "semver": "^5.3.0",
     "silent-error": "^1.0.0",
     "source-map": "^0.5.6",
-    "source-map-loader": "^0.1.5",
-    "sourcemap-istanbul-instrumenter-loader": "^0.2.0",
+    "source-map-loader": "^0.2.0",
+    "source-map-support": "^0.4.1",
     "style-loader": "^0.13.1",
     "stylus": "^0.54.5",
-    "stylus-loader": "^2.1.0",
+    "stylus-loader": "^3.0.1",
     "temp": "0.8.3",
-    "through": "^2.3.6",
-    "tslint": "^4.0.2",
-    "tslint-loader": "^3.3.0",
-    "typescript": "~2.0.3",
+    "typescript": "~2.4.2",
     "url-loader": "^0.5.7",
-    "walk-sync": "^0.2.6",
-    "webpack": "2.2.0-rc.3",
-    "webpack-dev-server": "2.2.0-rc.0",
-    "webpack-merge": "^0.14.0",
-    "webpack-sources": "^0.1.3",
-    "yam": "0.0.18",
-    "zone.js": "^0.7.2"
-  },
-  "ember-addon": {
-    "paths": [
-      "./packages/angular-cli/lib/addon"
-    ]
+    "walk-sync": "^0.3.1",
+    "webpack": "~3.4.1",
+    "webpack-dev-middleware": "^1.11.0",
+    "webpack-dev-server": "~2.5.1",
+    "webpack-merge": "^4.1.0",
+    "zone.js": "^0.8.14"
   },
   "devDependencies": {
-    "@types/chai": "^3.4.32",
+    "@angular/compiler": "^4.0.0",
+    "@angular/compiler-cli": "^4.0.0",
+    "@angular/core": "^4.0.0",
     "@types/chalk": "^0.4.28",
     "@types/common-tags": "^1.2.4",
-    "@types/denodeify": "^1.2.29",
+    "@types/denodeify": "^1.2.30",
     "@types/express": "^4.0.32",
-    "@types/fs-extra": "^0.0.31",
+    "@types/fs-extra": "^4.0.0",
     "@types/glob": "^5.0.29",
-    "@types/jasmine": "^2.2.32",
-    "@types/lodash": "^4.14.43",
+    "@types/jasmine": "2.5.45",
+    "@types/lodash": "~4.14.50",
+    "@types/minimist": "^1.2.0",
     "@types/mock-fs": "^3.6.30",
-    "@types/node": "^6.0.36",
-    "@types/request": "0.0.30",
-    "@types/rimraf": "0.0.25-alpha",
+    "@types/node": "^6.0.84",
+    "@types/request": "~2.0.0",
     "@types/semver": "^5.3.30",
     "@types/source-map": "^0.5.0",
-    "@types/webpack": "^1.12.22-alpha",
-    "chai": "^3.5.0",
+    "@types/webpack": "^3.0.5",
     "conventional-changelog": "^1.1.0",
-    "dtsgenerator": "^0.7.1",
-    "eslint": "^2.8.0",
+    "dtsgenerator": "^0.9.1",
+    "eslint": "^3.11.0",
     "express": "^4.14.0",
-    "jasmine": "^2.4.1",
-    "jasmine-spec-reporter": "^2.7.0",
+    "jasmine": "^2.6.0",
+    "jasmine-spec-reporter": "^4.1.0",
     "minimist": "^1.2.0",
-    "mocha": "^2.4.5",
-    "mock-fs": "3.10.0",
+    "mock-fs": "^4.0.0",
     "npm-run": "^4.1.0",
-    "npm-run-all": "^3.0.0",
-    "object-assign": "^4.0.1",
+    "npm-run-all": "^4.0.0",
     "request": "^2.74.0",
     "resolve-bin": "^0.4.0",
     "rewire": "^2.5.1",
-    "sinon": "^1.17.3",
+    "spdx-satisfies": "^0.1.3",
+    "tar": "^3.1.5",
+    "through": "^2.3.6",
     "tree-kill": "^1.0.0",
-    "ts-node": "^1.3.0"
+    "ts-node": "^3.2.0",
+    "tslint": "^5.1.0"
+  },
+  "optionalDependencies": {
+    "node-sass": "^4.3.0"
   }
 }
