@@ -2,26 +2,27 @@
   "name": "test",
   "license": "MIT",
   "dependencies": {
-    "@angular/common": "^2.3.1",
-    "@angular/compiler": "^2.3.1",
-    "@angular/compiler-cli": "^2.3.1",
-    "@angular/core": "^2.3.1",
-    "@angular/http": "^2.3.1",
-    "@angular/platform-browser": "^2.3.1",
-    "@angular/platform-browser-dynamic": "^2.3.1",
-    "@angular/platform-server": "^2.3.1",
-    "@angular/router": "^3.3.1",
+    "@angular/common": ">=4.0.0-beta <5.0.0",
+    "@angular/compiler": ">=4.0.0-beta <5.0.0",
+    "@angular/compiler-cli": ">=4.0.0-beta <5.0.0",
+    "@angular/core": ">=4.0.0-beta <5.0.0",
+    "@angular/http": ">=4.0.0-beta <5.0.0",
+    "@angular/platform-browser": ">=4.0.0-beta <5.0.0",
+    "@angular/platform-browser-dynamic": ">=4.0.0-beta <5.0.0",
+    "@angular/platform-server": ">=4.0.0-beta <5.0.0",
+    "@angular/router": ">=4.0.0-beta <5.0.0",
     "@ngtools/webpack": "0.0.0",
     "core-js": "^2.4.1",
-    "rxjs": "^5.0.1",
-    "zone.js": "^0.7.2"
+    "rxjs": "^5.4.2",
+    "zone.js": "^0.8.14"
   },
   "devDependencies": {
     "node-sass": "^3.7.0",
     "performance-now": "^0.2.0",
+    "preprocess-loader": "^0.2.2",
     "raw-loader": "^0.5.1",
     "sass-loader": "^3.2.0",
-    "typescript": "~2.0.3",
-    "webpack": "2.2.0-rc.3"
+    "typescript": "~2.4.2",
+    "webpack": "2.2.0"
   }
 }
