@@ -13,15 +13,15 @@
     "@angular/router": "^3.3.1",
     "@ngtools/webpack": "0.0.0",
     "core-js": "^2.4.1",
-    "rxjs": "^5.0.1",
-    "zone.js": "^0.7.2"
+    "rxjs": "^5.4.2",
+    "zone.js": "^0.8.14"
   },
   "devDependencies": {
     "node-sass": "^3.7.0",
     "performance-now": "^0.2.0",
     "raw-loader": "^0.5.1",
     "sass-loader": "^3.2.0",
-    "typescript": "~2.0.3",
-    "webpack": "2.2.0-rc.3"
+    "typescript": "~2.4.2",
+    "webpack": "2.2.0"
   }
 }
