@@ -1,35 +1,42 @@
-const fs = require('fs-extra');
-const ng = require('../helpers/ng');
-const existsSync = require('exists-sync');
-const expect = require('chai').expect;
-const forEach = require('lodash/forEach');
-const walkSync = require('walk-sync');
-const Blueprint = require('angular-cli/ember-cli/lib/models/blueprint');
-const path = require('path');
+// tslint:disable:max-line-length
+import * as fs from 'fs-extra';
+import * as path from 'path';
+import * as util from 'util';
+import { EOL } from 'os';
+import { forEach } from 'lodash';
+import { ng } from '../helpers';
+
 const tmp = require('../helpers/tmp');
+const walkSync = require('walk-sync');
+const Blueprint = require('@angular/cli/ember-cli/lib/models/blueprint');
+
 const root = process.cwd();
-const util = require('util');
-const EOL = require('os').EOL;
-const SilentError = require('silent-error');
 
 
 describe('Acceptance: ng new', function () {
-  beforeEach(function () {
-    return tmp.setup('./tmp').then(function () {
-      process.chdir('./tmp');
-    });
+  let originalTimeout: number;
+
+  beforeEach((done) => {
+    // Increase timeout for these tests only.
+    originalTimeout = jasmine.DEFAULT_TIMEOUT_INTERVAL;
+    jasmine.DEFAULT_TIMEOUT_INTERVAL = 10000;
+
+    spyOn(console, 'error');
+
+    tmp.setup('./tmp')
+      .then(() => process.chdir('./tmp'))
+      .then(() => done());
+  }, 10000);
+
+  afterEach((done) => {
+    jasmine.DEFAULT_TIMEOUT_INTERVAL = originalTimeout;
+    tmp.teardown('./tmp').then(() => done());
   });
 
-  afterEach(function () {
-    this.timeout(10000);
-
-    return tmp.teardown('./tmp');
-  });
-
-  function confirmBlueprintedForDir(dir) {
+  function confirmBlueprintedForDir(dir: string) {
     return function () {
       let blueprintPath = path.join(root, dir, 'files');
-      let expected = walkSync(blueprintPath);
+      let expected: string[] = walkSync(blueprintPath);
       let actual = walkSync('.').sort();
       let directory = path.basename(process.cwd());
 
@@ -38,13 +45,13 @@ describe('Acceptance: ng new', function () {
       });
 
       expected.forEach(function (file, index) {
-        expected[index] = file.replace(/__name__/g, 'angular-cli');
+        expected[index] = file.replace(/__name__/g, '@angular/cli');
       });
 
       expected.sort();
 
-      expect(directory).to.equal('foo');
-      expect(expected).to.deep.equal(
+      expect(directory).toBe('foo');
+      expect(expected).toEqual(
         actual,
         EOL + ' expected: ' + util.inspect(expected) + EOL + ' but got: ' + util.inspect(actual));
 
@@ -52,121 +59,145 @@ describe('Acceptance: ng new', function () {
   }
 
   function confirmBlueprinted() {
-    return confirmBlueprintedForDir('blueprints/ng2');
+    return confirmBlueprintedForDir('blueprints/ng');
   }
 
-  it('requires a valid name (!)', () => {
-    return ng(['new', '!', '--skip-npm', '--skip-git', '--inline-template'])
-      .then(() => { throw new Error(); }, () => {});
+  it('requires a valid name (!)', (done) => {
+    return ng(['new', '!', '--skip-install', '--skip-git', '--inline-template'])
+      .then(() => done.fail(), () => done());
   });
-  it('requires a valid name (abc-.)', () => {
-    return ng(['new', 'abc-.', '--skip-npm', '--skip-git', '--inline-template'])
-      .then(() => { throw new Error(); }, () => {});
+  it('requires a valid name (abc-.)', (done) => {
+    return ng(['new', 'abc-.', '--skip-install', '--skip-git', '--inline-template'])
+      .then(() => done.fail(), () => done());
   });
-  it('requires a valid name (abc-)', () => {
-    return ng(['new', 'abc-', '--skip-npm', '--skip-git', '--inline-template'])
-      .then(() => { throw new Error(); }, () => {});
+  it('requires a valid name (abc-)', (done) => {
+    return ng(['new', 'abc-', '--skip-install', '--skip-git', '--inline-template'])
+      .then(() => done.fail(), () => done());
   });
-  it('requires a valid name (abc-def-)', () => {
-    return ng(['new', 'abc-def-', '--skip-npm', '--skip-git', '--inline-template'])
-      .then(() => { throw new Error(); }, () => {});
+  it('requires a valid name (abc-def-)', (done) => {
+    return ng(['new', 'abc-def-', '--skip-install', '--skip-git', '--inline-template'])
+      .then(() => done.fail(), () => done());
   });
-  it('requires a valid name (abc-123)', () => {
-    return ng(['new', 'abc-123', '--skip-npm', '--skip-git', '--inline-template'])
-      .then(() => { throw new Error(); }, () => {});
+  it('requires a valid name (abc-123)', (done) => {
+    return ng(['new', 'abc-123', '--skip-install', '--skip-git', '--inline-template'])
+      .then(() => done.fail(), () => done());
   });
-  it('requires a valid name (abc)', () => {
-    return ng(['new', 'abc', '--skip-npm', '--skip-git', '--inline-template']);
+  it('requires a valid name (abc)', (done) => {
+    return ng(['new', 'abc', '--skip-install', '--skip-git', '--inline-template'])
+      .then(() => done(), () => done.fail());
   });
-  it('requires a valid name (abc-def)', () => {
-    return ng(['new', 'abc-def', '--skip-npm', '--skip-git', '--inline-template']);
-  });
-
-  it('ng new foo, where foo does not yet exist, works', function () {
-    return ng(['new', 'foo', '--skip-npm']).then(confirmBlueprinted);
+  it('requires a valid name (abc-def)', (done) => {
+    return ng(['new', 'abc-def', '--skip-install', '--skip-git', '--inline-template'])
+      .then(() => done(), () => done.fail());
   });
 
-  it('ng new with empty app does throw exception', function () {
-    expect(ng(['new', ''])).to.throw;
+  it('ng new foo, where foo does not yet exist, works', (done) => {
+    return ng(['new', 'foo', '--skip-install'])
+      .then(confirmBlueprinted)
+      .then(done, done.fail);
   });
 
-  it('ng new without app name does throw exception', function () {
-    expect(ng(['new', ''])).to.throw;
+  it('ng new with empty app does throw exception', (done) => {
+    return ng(['new', ''])
+      .then(() => done.fail(), () => done());
   });
 
-  it('ng new with app name creates new directory and has a dasherized package name', function () {
-    return ng(['new', 'FooApp', '--skip-npm', '--skip-git']).then(function () {
-      expect(!existsSync('FooApp'));
+  it('ng new without app name does throw exception', (done) => {
+    return ng(['new'])
+      .then(() => done.fail(), () => done());
+  });
+
+  it('ng new with app name creates new directory and has a dasherized package name', (done) => {
+    return ng(['new', 'FooApp', '--skip-install', '--skip-git']).then(() => {
+      expect(!fs.pathExistsSync('FooApp'));
 
       const pkgJson = JSON.parse(fs.readFileSync('package.json', 'utf8'));
-      expect(pkgJson.name).to.equal('foo-app');
-    });
+      expect(pkgJson.name).toBe('foo-app');
+    })
+    .then(done, done.fail);
   });
 
-  it('ng new has a .editorconfig file', function () {
-    return ng(['new', 'FooApp', '--skip-npm', '--skip-git']).then(function () {
-      expect(!existsSync('FooApp'));
+  it('ng new has a .editorconfig file', (done) => {
+    return ng(['new', 'FooApp', '--skip-install', '--skip-git']).then(() => {
+      expect(!fs.pathExistsSync('FooApp'));
 
       const editorConfig = fs.readFileSync('.editorconfig', 'utf8');
-      expect(editorConfig).to.exist;
-    });
+      expect(editorConfig).toBeDefined();
+    })
+    .then(done, done.fail);
   });
 
-  it('Cannot run ng new, inside of angular-cli project', function () {
-    return ng(['new', 'foo', '--skip-npm', '--skip-git'])
-      .then(function () {
-        return ng(['new', 'foo', '--skip-npm', '--skip-git']).then(() => {
-          throw new SilentError('Cannot run ng new, inside of angular-cli project should fail.');
+  it('Cannot run ng new, inside of Angular CLI project', (done) => {
+    return ng(['new', 'foo', '--skip-install', '--skip-git'])
+      .then(() => {
+        return ng(['new', 'foo', '--skip-install', '--skip-git']).then(() => {
+          done.fail();
         }, () => {
-          expect(!existsSync('foo'));
+          expect(!fs.pathExistsSync('foo'));
         });
       })
-      .then(confirmBlueprinted);
+      .then(confirmBlueprinted)
+      .then(done, done.fail);
   });
 
-  it('ng new without skip-git flag creates .git dir', function () {
-    return ng(['new', 'foo', '--skip-npm']).then(function () {
-      expect(existsSync('.git'));
-    });
+  it('ng new without skip-git flag creates .git dir', (done) => {
+    return ng(['new', 'foo', '--skip-install']).then(() => {
+      expect(fs.pathExistsSync('.git'));
+    })
+    .then(done, done.fail);
   });
 
-  it('ng new with --dry-run does not create new directory', function () {
-    return ng(['new', 'foo', '--dry-run']).then(function () {
+  it('ng new with --dry-run does not create new directory', (done) => {
+    return ng(['new', 'foo', '--dry-run']).then(() => {
       const cwd = process.cwd();
-      expect(cwd).to.not.match(/foo/, 'does not change cwd to foo in a dry run');
-      expect(!existsSync(path.join(cwd, 'foo')), 'does not create new directory');
-      expect(!existsSync(path.join(cwd, '.git')), 'does not create git in current directory');
-    });
+      expect(cwd).not.toMatch(/foo/, 'does not change cwd to foo in a dry run');
+      expect(fs.pathExistsSync(path.join(cwd, 'foo'))).toBe(false, 'does not create new directory');
+      expect(fs.pathExistsSync(path.join(cwd, '.git'))).toBe(false, 'does not create git in current directory');
+    })
+    .then(done, done.fail);
   });
 
-  it('ng new with --directory uses given directory name and has correct package name', function () {
-    return ng(['new', 'foo', '--skip-npm', '--skip-git', '--directory=bar'])
-      .then(function () {
+  it('ng new with --directory uses given directory name and has correct package name', (done) => {
+    return ng(['new', 'foo', '--skip-install', '--skip-git', '--directory=bar'])
+      .then(() => {
         const cwd = process.cwd();
-        expect(cwd).to.not.match(/foo/, 'does not use app name for directory name');
-        expect(!existsSync(path.join(cwd, 'foo')), 'does not create new directory with app name');
+        expect(cwd).not.toMatch(/foo/, 'does not use app name for directory name');
+        expect(fs.pathExistsSync(path.join(cwd, 'foo'))).toBe(false, 'does not create new directory with app name');
 
-        expect(cwd).to.match(/bar/, 'uses given directory name');
-        expect(existsSync(path.join(cwd, 'bar')), 'creates new directory with specified name');
+        expect(cwd).toMatch(/bar/, 'uses given directory name');
+        expect(fs.pathExistsSync(path.join(cwd, '..', 'bar'))).toBe(true, 'creates new directory with specified name');
 
         const pkgJson = JSON.parse(fs.readFileSync('package.json', 'utf8'));
-        expect(pkgJson.name).to.equal('foo', 'uses app name for package name');
-      });
+        expect(pkgJson.name).toBe('foo', 'uses app name for package name');
+      })
+      .then(done, done.fail);
   });
 
-  it('ng new --inline-template does not generate a template file', () => {
-    return ng(['new', 'foo', '--skip-npm', '--skip-git', '--inline-template'])
+  it('ng new --inline-template does not generate a template file', (done) => {
+    return ng(['new', 'foo', '--skip-install', '--skip-git', '--inline-template'])
       .then(() => {
         const templateFile = path.join('src', 'app', 'app.component.html');
-        expect(existsSync(templateFile)).to.equal(false);
-      });
+        expect(fs.pathExistsSync(templateFile)).toBe(false);
+      })
+      .then(done, done.fail);
   });
 
-  it('ng new --inline-style does not gener a style file', () => {
-    return ng(['new', 'foo', '--skip-npm', '--skip-git', '--inline-style'])
+  it('ng new --inline-style does not gener a style file', (done) => {
+    return ng(['new', 'foo', '--skip-install', '--skip-git', '--inline-style'])
       .then(() => {
         const styleFile = path.join('src', 'app', 'app.component.css');
-        expect(existsSync(styleFile)).to.equal(false);
-      });
+        expect(fs.pathExistsSync(styleFile)).toBe(false);
+      })
+      .then(done, done.fail);
   });
+
+  it('should skip spec files when passed --skip-tests', (done) => {
+    return ng(['new', 'foo', '--skip-install', '--skip-git', '--skip-tests'])
+      .then(() => {
+        const specFile = path.join('src', 'app', 'app.component.spec.ts');
+        expect(fs.pathExistsSync(specFile)).toBe(false);
+      })
+      .then(done, done.fail);
+  });
+
 });
