@@ -6,17 +6,19 @@ const path = require('path');
 const ts = require('typescript');
 
 
+Error.stackTraceLimit = Infinity;
+
 global.angularCliIsLocal = true;
-global.angularCliPackages = require('./packages');
+global.angularCliPackages = require('./packages').packages;
 
 const compilerOptions = JSON.parse(fs.readFileSync(path.join(__dirname, '../tsconfig.json')));
 
 const oldRequireTs = require.extensions['.ts'];
-require.extensions['.ts'] = function(m, filename) {
+require.extensions['.ts'] = function (m, filename) {
   // If we're in node module, either call the old hook or simply compile the
   // file without transpilation. We do not touch node_modules/**.
-  // We do touch `angular-cli` files anywhere though.
-  if (!filename.match(/angular-cli/) && filename.match(/node_modules/)) {
+  // We do touch `Angular CLI` files anywhere though.
+  if (!filename.match(/@angular\/cli/) && filename.match(/node_modules/)) {
     if (oldRequireTs) {
       return oldRequireTs(m, filename);
     }
@@ -46,17 +48,20 @@ require.extensions['.ts'] = function(m, filename) {
 
 // If we're running locally, meaning npm linked. This is basically "developer mode".
 if (!__dirname.match(new RegExp(`\\${path.sep}node_modules\\${path.sep}`))) {
-  const packages = require('./packages');
+  const packages = require('./packages').packages;
 
   // We mock the module loader so that we can fake our packages when running locally.
   const Module = require('module');
   const oldLoad = Module._load;
   Module._load = function (request, parent) {
+    if (request.match(/ts-node/) && parent && parent.id && parent.id.match(/karma/)) {
+      throw new Error();
+    }
     if (request in packages) {
       return oldLoad.call(this, packages[request].main, parent);
-    } else if (request.startsWith('angular-cli/')) {
+    } else if (request.startsWith('@angular/cli/')) {
       // We allow deep imports (for now).
-      // TODO: move tests to inside angular-cli package so they don't have to deep import.
+      // TODO: move tests to inside @angular/cli package so they don't have to deep import.
       const dir = path.dirname(parent.filename);
       const newRequest = path.relative(dir, path.join(__dirname, '../packages', request));
       return oldLoad.call(this, newRequest, parent);
