@@ -68,6 +68,10 @@ export class VirtualFileStats extends VirtualStats {
   set content(v: string) {
     this._content = v;
     this._mtime = new Date();
+    this._sourceFile = null;
+  }
+  setSourceFile(sourceFile: ts.SourceFile) {
+    this._sourceFile = sourceFile;
   }
   getSourceFile(languageVersion: ts.ScriptTarget, setParentNodes: boolean) {
     if (!this._sourceFile) {
@@ -91,10 +95,14 @@ export class WebpackCompilerHost implements ts.CompilerHost {
   private _delegate: ts.CompilerHost;
   private _files: {[path: string]: VirtualFileStats} = Object.create(null);
   private _directories: {[path: string]: VirtualDirStats} = Object.create(null);
-  private _changed = false;
+
+  private _changedFiles: {[path: string]: boolean} = Object.create(null);
+  private _changedDirs: {[path: string]: boolean} = Object.create(null);
 
   private _basePath: string;
   private _setParentNodes: boolean;
+
+  private _cache = false;
 
   constructor(private _options: ts.CompilerOptions, basePath: string) {
     this._setParentNodes = true;
@@ -109,11 +117,11 @@ export class WebpackCompilerHost implements ts.CompilerHost {
   private _resolve(path: string) {
     path = this._normalizePath(path);
     if (path[0] == '.') {
-      return join(this.getCurrentDirectory(), path);
+      return this._normalizePath(join(this.getCurrentDirectory(), path));
     } else if (path[0] == '/' || path.match(/^\w:\//)) {
       return path;
     } else {
-      return join(this._basePath, path);
+      return this._normalizePath(join(this._basePath, path));
     }
   }
 
@@ -123,54 +131,124 @@ export class WebpackCompilerHost implements ts.CompilerHost {
     let p = dirname(fileName);
     while (p && !this._directories[p]) {
       this._directories[p] = new VirtualDirStats(p);
+      this._changedDirs[p] = true;
       p = dirname(p);
     }
 
-    this._changed = true;
+    this._changedFiles[fileName] = true;
+  }
+
+  get dirty() {
+    return Object.keys(this._changedFiles).length > 0;
+  }
+
+  enableCaching() {
+    this._cache = true;
   }
 
   populateWebpackResolver(resolver: any) {
     const fs = resolver.fileSystem;
-    if (!this._changed) {
+    if (!this.dirty) {
       return;
     }
 
-    const isWindows = process.platform.startsWith('win');
-    for (const fileName of Object.keys(this._files)) {
-      const stats = this._files[fileName];
-      // If we're on windows, we need to populate with the proper path separator.
-      const path = isWindows ? fileName.replace(/\//g, '\\') : fileName;
-      fs._statStorage.data[path] = [null, stats];
-      fs._readFileStorage.data[path] = [null, stats.content];
+    /**
+     * storageDataSetter is a temporary hack to address these two issues:
+     * - https://github.com/angular/angular-cli/issues/7113
+     * - https://github.com/angular/angular-cli/issues/7136
+     *
+     * This way we set values correctly in both a Map (enhanced-resove>=3.4.0) and
+     * object (enhanced-resolve >= 3.1.0 <3.4.0).
+     *
+     * The right solution is to create a virtual filesystem by decorating the filesystem,
+     * instead of injecting data into the private cache of the filesystem.
+     *
+     * Doing it the right way should fix other related bugs, but meanwhile we hack it since:
+     * - it's affecting a lot of users.
+     * - the real solution is non-trivial.
+     */
+    function storageDataSetter(data: Map<string, any> | {[k: string]: any}, k: string, v: any) {
+
+      if (data instanceof Map) {
+        data.set(k, v);
+      } else {
+        data[k] = v;
+      }
     }
-    for (const dirName of Object.keys(this._directories)) {
+
+
+
+    const isWindows = process.platform.startsWith('win');
+    for (const fileName of this.getChangedFilePaths()) {
+      const stats = this._files[fileName];
+      if (stats) {
+        // If we're on windows, we need to populate with the proper path separator.
+        const path = isWindows ? fileName.replace(/\//g, '\\') : fileName;
+        // fs._statStorage.data[path] = [null, stats];
+        // fs._readFileStorage.data[path] = [null, stats.content];
+        storageDataSetter(fs._statStorage.data, path, [null, stats]);
+        storageDataSetter(fs._readFileStorage.data, path, [null, stats.content]);
+      } else {
+        // Support removing files as well.
+        const path = isWindows ? fileName.replace(/\//g, '\\') : fileName;
+        // fs._statStorage.data[path] = [new Error(), null];
+        // fs._readFileStorage.data[path] = [new Error(), null];
+        storageDataSetter(fs._statStorage.data, path, [new Error(), null]);
+        storageDataSetter(fs._readFileStorage.data, path, [new Error(), null]);
+      }
+    }
+    for (const dirName of Object.keys(this._changedDirs)) {
       const stats = this._directories[dirName];
       const dirs = this.getDirectories(dirName);
       const files = this.getFiles(dirName);
       // If we're on windows, we need to populate with the proper path separator.
       const path = isWindows ? dirName.replace(/\//g, '\\') : dirName;
-      fs._statStorage.data[path] = [null, stats];
-      fs._readdirStorage.data[path] = [null, files.concat(dirs)];
+      // fs._statStorage.data[path] = [null, stats];
+      // fs._readdirStorage.data[path] = [null, files.concat(dirs)];
+      storageDataSetter(fs._statStorage.data, path, [null, stats]);
+      storageDataSetter(fs._readFileStorage.data, path, [null, files.concat(dirs)]);
     }
+  }
 
-    this._changed = false;
+  resetChangedFileTracker() {
+    this._changedFiles = Object.create(null);
+    this._changedDirs = Object.create(null);
+  }
+  getChangedFilePaths(): string[] {
+    return Object.keys(this._changedFiles);
+  }
+
+  invalidate(fileName: string): void {
+    fileName = this._resolve(fileName);
+    if (fileName in this._files) {
+      this._files[fileName] = null;
+      this._changedFiles[fileName] = true;
+    }
   }
 
   fileExists(fileName: string): boolean {
     fileName = this._resolve(fileName);
-    return fileName in this._files || this._delegate.fileExists(fileName);
+    return this._files[fileName] != null || this._delegate.fileExists(fileName);
   }
 
   readFile(fileName: string): string {
     fileName = this._resolve(fileName);
-    return (fileName in this._files)
-         ? this._files[fileName].content
-         : this._delegate.readFile(fileName);
+    if (this._files[fileName] == null) {
+      const result = this._delegate.readFile(fileName);
+      if (result !== undefined && this._cache) {
+        this._setFileContent(fileName, result);
+        return result;
+      } else {
+        return result;
+      }
+    }
+    return this._files[fileName].content;
   }
 
   directoryExists(directoryName: string): boolean {
     directoryName = this._resolve(directoryName);
-    return (directoryName in this._directories) || this._delegate.directoryExists(directoryName);
+    return (this._directories[directoryName] != null)
+        || this._delegate.directoryExists(directoryName);
   }
 
   getFiles(path: string): string[] {
@@ -195,11 +273,14 @@ export class WebpackCompilerHost implements ts.CompilerHost {
     return delegated.concat(subdirs);
   }
 
-  getSourceFile(fileName: string, languageVersion: ts.ScriptTarget, onError?: OnErrorFn) {
+  getSourceFile(fileName: string, languageVersion: ts.ScriptTarget, _onError?: OnErrorFn) {
     fileName = this._resolve(fileName);
 
-    if (!(fileName in this._files)) {
-      return this._delegate.getSourceFile(fileName, languageVersion, onError);
+    if (this._files[fileName] == null) {
+      const content = this.readFile(fileName);
+      if (!this._cache) {
+        return ts.createSourceFile(fileName, content, languageVersion, this._setParentNodes);
+      }
     }
 
     return this._files[fileName].getSourceFile(languageVersion, this._setParentNodes);
@@ -216,8 +297,8 @@ export class WebpackCompilerHost implements ts.CompilerHost {
   // This is due to typescript CompilerHost interface being weird on writeFile. This shuts down
   // typings in WebStorm.
   get writeFile() {
-    return (fileName: string, data: string, writeByteOrderMark: boolean,
-            onError?: (message: string) => void, sourceFiles?: ts.SourceFile[]): void => {
+    return (fileName: string, data: string, _writeByteOrderMark: boolean,
+            _onError?: (message: string) => void, _sourceFiles?: ts.SourceFile[]): void => {
       fileName = this._resolve(fileName);
       this._setFileContent(fileName, data);
     };
