@@ -1,22 +1,23 @@
 import * as path from 'path';
 import * as ts from 'typescript';
-import {Request, ResolverPlugin, Callback, Tapable} from './webpack';
+import {
+  ResolverPlugin,
+  Callback,
+  Tapable,
+  NormalModuleFactory,
+  NormalModuleFactoryRequest,
+} from './webpack';
 
 
 const ModulesInRootPlugin: new (a: string, b: string, c: string) => ResolverPlugin
   = require('enhanced-resolve/lib/ModulesInRootPlugin');
 
-interface CreateInnerCallback {
-  (callback: Callback<any>,
-   options: Callback<any>,
-   message?: string,
-   messageOptional?: string): Callback<any>;
+export interface Mapping {
+  onlyModule: boolean;
+  alias: string;
+  aliasPattern: RegExp;
+  target: string;
 }
-
-const createInnerCallback: CreateInnerCallback
-  = require('enhanced-resolve/lib/createInnerCallback');
-const getInnerRequest: (resolver: ResolverPlugin, request: Request) => string
-  = require('enhanced-resolve/lib/getInnerRequest');
 
 
 function escapeRegExp(str: string): string {
@@ -25,12 +26,14 @@ function escapeRegExp(str: string): string {
 
 
 export interface PathsPluginOptions {
+  nmf: NormalModuleFactory;
   tsConfigPath: string;
   compilerOptions?: ts.CompilerOptions;
   compilerHost?: ts.CompilerHost;
 }
 
 export class PathsPlugin implements Tapable {
+  private _nmf: NormalModuleFactory;
   private _tsConfigPath: string;
   private _compilerOptions: ts.CompilerOptions;
   private _host: ts.CompilerHost;
@@ -38,7 +41,7 @@ export class PathsPlugin implements Tapable {
   source: string;
   target: string;
 
-  private mappings: any;
+  private _mappings: Mapping[];
 
   private _absoluteBaseUrl: string;
 
@@ -54,7 +57,7 @@ export class PathsPlugin implements Tapable {
     if (tsConfig.error) {
       throw tsConfig.error;
     }
-    return tsConfig.config;
+    return tsConfig.config.compilerOptions;
   }
 
   constructor(options: PathsPluginOptions) {
@@ -76,6 +79,7 @@ export class PathsPlugin implements Tapable {
       this._host = ts.createCompilerHost(this._compilerOptions, false);
     }
 
+    this._nmf = options.nmf;
     this.source = 'described-resolve';
     this.target = 'resolve';
 
@@ -84,7 +88,7 @@ export class PathsPlugin implements Tapable {
       this._compilerOptions.baseUrl || '.'
     );
 
-    this.mappings = [];
+    this._mappings = [];
     let paths = this._compilerOptions.paths || {};
     Object.keys(paths).forEach(alias => {
       let onlyModule = alias.indexOf('*') === -1;
@@ -99,7 +103,7 @@ export class PathsPlugin implements Tapable {
           aliasPattern = new RegExp(`^${withStarCapturing}`);
         }
 
-        this.mappings.push({
+        this._mappings.push({
           onlyModule,
           alias,
           aliasPattern,
@@ -110,65 +114,41 @@ export class PathsPlugin implements Tapable {
   }
 
   apply(resolver: ResolverPlugin): void {
-    let { baseUrl } = this._compilerOptions;
+    let baseUrl = this._compilerOptions.baseUrl || '.';
 
     if (baseUrl) {
       resolver.apply(new ModulesInRootPlugin('module', this._absoluteBaseUrl, 'resolve'));
     }
 
-    this.mappings.forEach((mapping: any) => {
-      resolver.plugin(this.source, this.createPlugin(resolver, mapping));
-    });
-  }
+    this._nmf.plugin('before-resolve', (request: NormalModuleFactoryRequest,
+                                        callback: Callback<any>) => {
+      for (let mapping of this._mappings) {
+        const match = request.request.match(mapping.aliasPattern);
+        if (!match) { continue; }
 
-  resolve(resolver: ResolverPlugin, mapping: any, request: any, callback: Callback<any>): any {
-    let innerRequest = getInnerRequest(resolver, request);
-    if (!innerRequest) {
-      return callback();
-    }
+        let newRequestStr = mapping.target;
+        if (!mapping.onlyModule) {
+          newRequestStr = newRequestStr.replace('*', match[1]);
+        }
 
-    let match = innerRequest.match(mapping.aliasPattern);
-    if (!match) {
-      return callback();
-    }
+        const moduleResolver: ts.ResolvedModuleWithFailedLookupLocations =
+          ts.nodeModuleNameResolver(
+            newRequestStr,
+            this._absoluteBaseUrl,
+            this._compilerOptions,
+            this._host
+          );
+        const moduleFilePath = moduleResolver.resolvedModule ?
+          moduleResolver.resolvedModule.resolvedFileName : '';
 
-    let newRequestStr = mapping.target;
-    if (!mapping.onlyModule) {
-      newRequestStr = newRequestStr.replace('*', match[1]);
-    }
-    if (newRequestStr[0] === '.') {
-      newRequestStr = path.resolve(this._absoluteBaseUrl, newRequestStr);
-    }
-
-    let newRequest = Object.assign({}, request, {
-      request: newRequestStr
-    }) as Request;
-
-    return resolver.doResolve(
-      this.target,
-      newRequest,
-      `aliased with mapping '${innerRequest}': '${mapping.alias}' to '${newRequestStr}'`,
-      createInnerCallback(
-        function(err, result) {
-          if (arguments.length > 0) {
-            return callback(err, result);
-          }
-
-          // don't allow other aliasing or raw request
-          callback(null, null);
-        },
-        callback
-      )
-    );
-  }
-
-  createPlugin(resolver: ResolverPlugin, mapping: any): any {
-    return (request: any, callback: Callback<any>) => {
-      try {
-        this.resolve(resolver, mapping, request, callback);
-      } catch (err) {
-        callback(err);
+        if (moduleFilePath) {
+          return callback(null, Object.assign({}, request, {
+            request: moduleFilePath.includes('.d.ts') ? newRequestStr : moduleFilePath
+          }));
+        }
       }
-    };
+
+      return callback(null, request);
+    });
   }
 }
