@@ -12,7 +12,7 @@ export interface TranspileOutput {
 }
 
 
-function resolve(filePath: string, host: ts.CompilerHost, program: ts.Program) {
+function resolve(filePath: string, _host: ts.CompilerHost, program: ts.Program) {
   if (path.isAbsolute(filePath)) {
     return filePath;
   }
@@ -30,24 +30,28 @@ export class TypeScriptFileRefactor {
   private _sourceFile: ts.SourceFile;
   private _sourceString: any;
   private _sourceText: string;
-  private _changed: boolean = false;
+  private _changed = false;
 
   get fileName() { return this._fileName; }
   get sourceFile() { return this._sourceFile; }
   get sourceText() { return this._sourceString.toString(); }
 
   constructor(fileName: string,
-              private _host: ts.CompilerHost,
-              private _program?: ts.Program) {
+              _host: ts.CompilerHost,
+              private _program?: ts.Program,
+              source?: string | null) {
     fileName = resolve(fileName, _host, _program).replace(/\\/g, '/');
     this._fileName = fileName;
     if (_program) {
-      this._sourceFile = _program.getSourceFile(fileName);
+      if (source) {
+        this._sourceFile = ts.createSourceFile(fileName, source, ts.ScriptTarget.Latest, true);
+      } else {
+        this._sourceFile = _program.getSourceFile(fileName);
+      }
     }
     if (!this._sourceFile) {
-      this._program = null;
-      this._sourceFile = ts.createSourceFile(fileName, _host.readFile(fileName),
-        ts.ScriptTarget.Latest);
+      this._sourceFile = ts.createSourceFile(fileName, source || _host.readFile(fileName),
+        ts.ScriptTarget.Latest, true);
     }
     this._sourceText = this._sourceFile.getFullText(this._sourceFile);
     this._sourceString = new MagicString(this._sourceText);
@@ -56,16 +60,19 @@ export class TypeScriptFileRefactor {
   /**
    * Collates the diagnostic messages for the current source file
    */
-  getDiagnostics(): ts.Diagnostic[] {
+  getDiagnostics(typeCheck = true): ts.Diagnostic[] {
     if (!this._program) {
       return [];
     }
-    let diagnostics: ts.Diagnostic[] = this._program.getSyntacticDiagnostics(this._sourceFile)
-                              .concat(this._program.getSemanticDiagnostics(this._sourceFile));
+    let diagnostics: ts.Diagnostic[] = [];
     // only concat the declaration diagnostics if the tsconfig config sets it to true.
     if (this._program.getCompilerOptions().declaration == true) {
       diagnostics = diagnostics.concat(this._program.getDeclarationDiagnostics(this._sourceFile));
     }
+    diagnostics = diagnostics.concat(
+      this._program.getSyntacticDiagnostics(this._sourceFile),
+      typeCheck ? this._program.getSemanticDiagnostics(this._sourceFile) : []);
+
     return diagnostics;
   }
 
@@ -80,7 +87,7 @@ export class TypeScriptFileRefactor {
   findAstNodes(node: ts.Node | null,
                kind: ts.SyntaxKind,
                recursive = false,
-               max: number = Infinity): ts.Node[] {
+               max = Infinity): ts.Node[] {
     if (max == 0) {
       return [];
     }
@@ -117,8 +124,19 @@ export class TypeScriptFileRefactor {
     return arr;
   }
 
+  findFirstAstNode(node: ts.Node | null, kind: ts.SyntaxKind): ts.Node | null {
+    return this.findAstNodes(node, kind, false, 1)[0] || null;
+  }
+
   appendAfter(node: ts.Node, text: string): void {
-    this._sourceString.insertRight(node.getEnd(), text);
+    this._sourceString.appendRight(node.getEnd(), text);
+  }
+  append(node: ts.Node, text: string): void {
+    this._sourceString.appendLeft(node.getEnd(), text);
+  }
+
+  prependBefore(node: ts.Node, text: string) {
+    this._sourceString.appendLeft(node.getStart(), text);
   }
 
   insertImport(symbolName: string, modulePath: string): void {
@@ -170,7 +188,7 @@ export class TypeScriptFileRefactor {
   }
 
   removeNodes(...nodes: ts.Node[]) {
-    nodes.forEach(node => this.removeNode(node));
+    nodes.forEach(node => node && this.removeNode(node));
   }
 
   replaceNode(node: ts.Node, replacement: string) {
@@ -178,7 +196,7 @@ export class TypeScriptFileRefactor {
     this._sourceString.overwrite(node.getStart(this._sourceFile),
                                  node.getEnd(),
                                  replacement,
-                                 replaceSymbolName);
+                                 { storeName: replaceSymbolName });
     this._changed = true;
   }
 
