@@ -1,4 +1,4 @@
-import 'reflect-metadata';
+import 'core-js/es7/reflect';
 import {platformBrowserDynamic} from '@angular/platform-browser-dynamic';
 import {AppModule} from './app.module';
 
