@@ -1,5 +1,10 @@
 const ngToolsWebpack = require('@ngtools/webpack');
 
+const flags = require('./webpack.flags.json');
+
+const preprocessLoader = 'preprocess-loader' + (flags.DEBUG ? '?+DEBUG' : '');
+
+
 module.exports = {
   resolve: {
     extensions: ['.ts', '.js']
@@ -15,10 +20,14 @@ module.exports = {
   ],
   module: {
     loaders: [
-      { test: /\.scss$/, loaders: ['raw-loader', 'sass-loader'] },
+      { test: /\.scss$/, loaders: ['raw-loader', 'sass-loader', preprocessLoader] },
       { test: /\.css$/, loader: 'raw-loader' },
-      { test: /\.html$/, loader: 'raw-loader' },
-      { test: /\.ts$/, loader: '@ngtools/webpack' }
+      { test: /\.html$/, loaders: ['raw-loader', preprocessLoader] },
+      // Use preprocess to remove DEBUG only code.
+      { test: /\.ts$/, use: [
+        { loader: '@ngtools/webpack' },
+        { loader: preprocessLoader }
+      ] }
     ]
   },
   devServer: {
