@@ -3,24 +3,16 @@ import {getGlobalVariable} from '../utils/env';
 import {npm} from '../utils/process';
 import {updateJsonFile} from '../utils/project';
 
-const packages = require('../../../lib/packages');
+const packages = require('../../../lib/packages').packages;
 
 export default function() {
   const argv = getGlobalVariable('argv');
 
-  return npm('run', 'build')
-    .then(() => console.log('Updating package.json from dist...'))
-    .then(() => Promise.all(Object.keys(packages).map(pkgName => {
-      return updateJsonFile(join(packages[pkgName].dist, 'package.json'), json => {
-        Object.keys(packages).forEach(pkgName => {
-          if (json['dependencies'] && pkgName in json['dependencies']) {
-            json['dependencies'][pkgName] = packages[pkgName].dist;
-          } else if (json['devDependencies'] && pkgName in json['devDependencies']) {
-            json['devDependencies'][pkgName] = packages[pkgName].dist;
-          }
-        });
-      });
-    })))
+  if (argv.nobuild) {
+    return;
+  }
+
+  return npm('run', 'build', '--', '--local')
     .then(() => {
       if (!argv.nightly && !argv['ng-sha']) {
         return;
