@@ -1,9 +1,9 @@
 import {join} from 'path';
-import {npm, exec} from '../utils/process';
+import {silentNpm, exec} from '../utils/process';
 import {updateJsonFile} from '../utils/project';
 import {getGlobalVariable} from '../utils/env';
 
-const packages = require('../../../lib/packages');
+const packages = require('../../../lib/packages').packages;
 
 
 export default function () {
@@ -14,7 +14,7 @@ export default function () {
         return;
       }
 
-      const distAngularCli = join(__dirname, '../../../dist/angular-cli');
+      const distAngularCli = packages['@angular/cli'].dist;
       const oldCwd = process.cwd();
       process.chdir(distAngularCli);
 
@@ -33,7 +33,8 @@ export default function () {
           }
         });
       }))
-      .then(() => npm('link'))
+      // npm link is very noisy on the older version used in Circle CI.
+      .then(() => silentNpm('link'))
       .then(() => process.chdir(oldCwd));
     })
     .then(() => exec(process.platform.startsWith('win') ? 'where' : 'which', 'ng'));
