@@ -1,14 +1,20 @@
 import { request } from '../../utils/http';
 import { killAllProcesses } from '../../utils/process';
 import { ngServe } from '../../utils/project';
+import { getGlobalVariable } from '../../utils/env';
 
 
 export default function() {
+  // Skip this in Appveyor tests.
+  if (getGlobalVariable('argv').appveyor) {
+    return Promise.resolve();
+  }
+
   return Promise.resolve()
     .then(() => ngServe('--ssl', 'true'))
     .then(() => request('https://localhost:4200/'))
     .then(body => {
-      if (!body.match(/<app-root>Loading...<\/app-root>/)) {
+      if (!body.match(/<app-root><\/app-root>/)) {
         throw new Error('Response does not match expected value.');
       }
     })
