@@ -4,7 +4,7 @@ import {ng} from '../../utils/process';
 
 export default function() {
   return ng('version')
-    .then(() => deleteFile('angular-cli.json'))
-    // doesn't fail on a project with missing angular-cli.json
+    .then(() => deleteFile('.angular-cli.json'))
+    // doesn't fail on a project with missing .angular-cli.json
     .then(() => ng('version'));
 }
