@@ -2,10 +2,16 @@ import { writeMultipleFiles } from '../../utils/fs';
 import { ng } from '../../utils/process';
 import { updateJsonFile } from '../../utils/project';
 import { expectToFail } from '../../utils/utils';
+import { getGlobalVariable } from '../../utils/env';
 import { stripIndent } from 'common-tags';
 
 // Make sure asset files are served
 export default function () {
+  // Skip this in Appveyor tests.
+  if (getGlobalVariable('argv').appveyor) {
+    return Promise.resolve();
+  }
+
   return Promise.resolve()
     .then(() => writeMultipleFiles({
       'src/assets/file.txt': 'assets-folder-content',
@@ -33,15 +39,15 @@ export default function () {
         });
       `
     }))
-    // Test failure condition (no assets in angular-cli.json)
-    .then(() => updateJsonFile('angular-cli.json', configJson => {
+    // Test failure condition (no assets in .angular-cli.json)
+    .then(() => updateJsonFile('.angular-cli.json', configJson => {
       const app = configJson['apps'][0];
       app['assets'] = [];
     }))
     .then(() => expectToFail(() => ng('test', '--single-run'),
-      'Should fail because the assets to serve were not in the angular-cli config'))
+      'Should fail because the assets to serve were not in the Angular CLI config'))
     // Test passing condition (assets are included)
-    .then(() => updateJsonFile('angular-cli.json', configJson => {
+    .then(() => updateJsonFile('.angular-cli.json', configJson => {
       const app = configJson['apps'][0];
       app['assets'] = ['assets', 'file.txt'];
     }))
