@@ -2,9 +2,15 @@ import { writeMultipleFiles } from '../../utils/fs';
 import { ng } from '../../utils/process';
 import { updateJsonFile } from '../../utils/project';
 import { expectToFail } from '../../utils/utils';
+import { getGlobalVariable } from '../../utils/env';
 import { stripIndent } from 'common-tags';
 
 export default function () {
+  // Skip this in Appveyor tests.
+  if (getGlobalVariable('argv').appveyor) {
+    return Promise.resolve();
+  }
+
   return Promise.resolve()
     .then(() => ng('test', '--watch=false'))
     // prepare global scripts test files
@@ -29,10 +35,11 @@ export default function () {
         import { AppComponent } from './app.component';
 
         describe('AppComponent', () => {
-          beforeEach(() => {
-            TestBed.configureTestingModule({ declarations: [ AppComponent ] });
-            TestBed.compileComponents();
-          });
+          beforeEach(async(() => {
+            TestBed.configureTestingModule({
+              declarations: [ AppComponent ]
+            }).compileComponents();
+          }));
 
           it('should have access to string-script.js', async(() => {
             let app = TestBed.createComponent(AppComponent).debugElement.componentInstance;
@@ -58,7 +65,7 @@ export default function () {
     }))
     // should fail because the global scripts were not added to scripts array
     .then(() => expectToFail(() => ng('test', '--single-run')))
-    .then(() => updateJsonFile('angular-cli.json', configJson => {
+    .then(() => updateJsonFile('.angular-cli.json', configJson => {
       const app = configJson['apps'][0];
       app['scripts'] = [
         'string-script.js',
