@@ -1,4 +1,4 @@
-import * as fs from 'fs';
+import * as fs from 'fs-extra';
 import {dirname} from 'path';
 import {stripIndents} from 'common-tags';
 
@@ -15,9 +15,9 @@ export function readFile(fileName: string) {
   });
 }
 
-export function writeFile(fileName: string, content: string) {
+export function writeFile(fileName: string, content: string, options?: any) {
   return new Promise<void>((resolve, reject) => {
-    fs.writeFile(fileName, content, (err: any) => {
+    fs.writeFile(fileName, content, options, (err: any) => {
       if (err) {
         reject(err);
       } else {
@@ -41,6 +41,19 @@ export function deleteFile(path: string) {
 }
 
 
+export function rimraf(path: string) {
+  return new Promise<void>((resolve, reject) => {
+    fs.remove(path, (err?: any) => {
+      if (err) {
+        reject(err);
+      } else {
+        resolve();
+      }
+    });
+  });
+}
+
+
 export function moveFile(from: string, to: string) {
   return new Promise<void>((resolve, reject) => {
     fs.rename(from, to, (err) => {
@@ -53,12 +66,25 @@ export function moveFile(from: string, to: string) {
   });
 }
 
+
+export function symlinkFile(from: string, to: string, type?: string) {
+  return new Promise<void>((resolve, reject) => {
+    fs.symlink(from, to, type, (err) => {
+      if (err) {
+        reject(err);
+      } else {
+        resolve();
+      }
+    });
+  });
+}
+
 export function createDir(path: string) {
-  _recursiveMkDir(path);
+  return _recursiveMkDir(path);
 }
 
 
-function _recursiveMkDir(path: string) {
+function _recursiveMkDir(path: string): Promise<void> {
   if (fs.existsSync(path)) {
     return Promise.resolve();
   } else {
@@ -71,37 +97,61 @@ export function copyFile(from: string, to: string) {
   return _recursiveMkDir(dirname(to))
     .then(() => new Promise((resolve, reject) => {
       const rd = fs.createReadStream(from);
-      rd.on('error', (err) => reject(err));
+      rd.on('error', (err: Error) => reject(err));
 
       const wr = fs.createWriteStream(to);
-      wr.on('error', (err) => reject(err));
-      wr.on('close', (ex) => resolve());
+      wr.on('error', (err: Error) => reject(err));
+      wr.on('close', () => resolve());
 
       rd.pipe(wr);
     }));
 }
 
 
-export function writeMultipleFiles(fs: any) {
-  return Object.keys(fs)
-    .reduce((previous, curr) => {
-      return previous.then(() => writeFile(curr, fs[curr]));
-    }, Promise.resolve());
+export function writeMultipleFiles(fs: { [path: string]: string }) {
+  return Promise.all(Object.keys(fs).map(fileName => writeFile(fileName, fs[fileName])));
 }
 
 
-export function replaceInFile(filePath: string, match: string, replacement: string);
-export function replaceInFile(filePath: string, match: RegExp, replacement: string) {
+export function replaceInFile(filePath: string, match: RegExp | string, replacement: string) {
   return readFile(filePath)
     .then((content: string) => writeFile(filePath, content.replace(match, replacement)));
 }
 
 
-export function appendToFile(filePath: string, text: string) {
+export function appendToFile(filePath: string, text: string, options?: any) {
   return readFile(filePath)
-    .then((content: string) => writeFile(filePath, content.concat(text)));
+    .then((content: string) => writeFile(filePath, content.concat(text), options));
 }
 
+
+export function prependToFile(filePath: string, text: string, options?: any) {
+  return readFile(filePath)
+    .then((content: string) => writeFile(filePath, text.concat(content), options));
+}
+
+
+export function expectFileMatchToExist(dir: string, regex: RegExp) {
+  return new Promise((resolve, reject) => {
+    const [fileName] = fs.readdirSync(dir).filter(name => name.match(regex));
+    if (!fileName) {
+      reject(new Error(`File ${regex} was expected to exist but not found...`));
+    }
+    resolve(fileName);
+  });
+}
+
+export function expectFileNotToExist(fileName: string) {
+  return new Promise((resolve, reject) => {
+    fs.exists(fileName, (exist) => {
+      if (exist) {
+        reject(new Error(`File ${fileName} was expected not to exist but found...`));
+      } else {
+        resolve();
+      }
+    });
+  });
+}
 
 export function expectFileToExist(fileName: string) {
   return new Promise((resolve, reject) => {
