@@ -1,15 +1,13 @@
 'use strict';
 
-var fs = require('fs-extra');
-var existsSync = require('exists-sync');
-var Promise = require('angular-cli/ember-cli/lib/ext/promise');
-var remove = Promise.denodeify(fs.remove);
-var root = process.cwd();
+const fs = require('fs-extra');
+
+const root = process.cwd();
 
 module.exports.setup = function (path) {
   process.chdir(root);
 
-  return remove(path).then(function () {
+  return fs.remove(path).then(function () {
     fs.mkdirsSync(path);
   });
 };
@@ -17,8 +15,8 @@ module.exports.setup = function (path) {
 module.exports.teardown = function (path) {
   process.chdir(root);
 
-  if (existsSync(path)) {
-    return remove(path);
+  if (fs.pathExistsSync(path)) {
+    return fs.remove(path);
   } else {
     return Promise.resolve();
   }
